fix(#692): improves tests to make it easy to make parts reusable
